Revert "enhancement(palette): position (left, top, bottom, right) is the same as space (ml, mx, etc) now, and we added % (#4445)"
This reverts commit e20995ef9030412e86c1a2557613fd3ef22e94be.
