+++ CHANGELOG.yml
@@ -13,7 +13,6 @@ upcoming:
     - hooks up inquiry make offer modal to web view checkout flow - lily
     - Convert cents to dollars in market stats - pepopowitz
     - Dedupe currencies in auction results - pepopowitz
-    - Add better types in space and position props in palette - pavlos
     - update support email - brian
     - scroll to the top after applying a filter on auction results- mounir
     - Market stats formatting fixes - brian
