+++ package.json
@@ -222,7 +222,7 @@
     "tslint": "5.14.0",
     "tslint-config-prettier": "1.18.0",
     "tslint-plugin-relay": "0.0.2",
-    "typescript": "4.2.0-beta",
+    "typescript": "^4.1.2",
     "typescript-styled-plugin": "0.13.0",
     "vscode-apollo-relay": "1.5.0",
     "yaml": "1.9.2",
