+++ src/palette/elements/Box/Box.tsx
@@ -1,5 +1,4 @@
-import { paletteSpace, PositionProps, SpaceProps } from "palette/helpers"
-import { palettePosition } from "palette/helpers/styled-system/position"
+import { paletteSpace, SpaceProps } from "palette/helpers"
 import { View, ViewProps } from "react-native"
 import styled from "styled-components/native"
 import {
@@ -14,6 +13,8 @@ import {
   FlexboxProps,
   layout,
   LayoutProps,
+  position,
+  PositionProps,
   textAlign,
   TextAlignProps,
 } from "styled-system"
@@ -32,7 +33,7 @@ export interface BoxProps
 /**
  * All the system functions for Box
  */
-export const boxMixin = compose(background, border, color, flexbox, layout, palettePosition, paletteSpace, textAlign)
+export const boxMixin = compose(background, border, color, flexbox, layout, position, paletteSpace, textAlign)
 
 /**
  * Box is just a `View` with common styled-system props.
