+++ src/palette/elements/Cards/CardTag.tsx
@@ -1,7 +1,7 @@
-import { PositionProps } from "palette"
 import React from "react"
 import { View } from "react-native"
 import { StyleProp, ViewStyle } from "react-native"
+import { PositionProps } from "styled-system"
 
 import { Sans } from ".."
 import { color } from "../../Theme"
