+++ src/palette/helpers/types.ts
@@ -1,33 +1,10 @@
 import { SpacingUnit as NumberSpacingUnit } from "@artsy/palette-tokens"
-import {
-  BottomProps,
-  LeftProps,
-  PositionProps as OGPositionProps,
-  RightProps,
-  SpaceProps as OGSpaceProps,
-  TopProps,
-  ZIndexProps,
-} from "styled-system"
+import { SpaceProps as OGSpaceProps } from "styled-system"
 
 type ToString<N extends number> = `${N}`
 type ToNegative<S extends string> = `-${S}`
 
-export type Percentage = `${number}%`
-
-export type SpacingUnit =
-  | number // pixels
-  | ToString<NumberSpacingUnit> // from our design system
-  | ToNegative<ToString<NumberSpacingUnit>> // from our design system
-  | "auto" // stays as is
-  | Percentage // stays as is
+// number for pixels, NumberSpacingUnit string for our design system
+export type SpacingUnit = number | ToString<NumberSpacingUnit> | ToNegative<ToString<NumberSpacingUnit>> | "auto"
 
 export type SpaceProps = OGSpaceProps<{}, SpacingUnit>
-
-export interface PositionProps
-  extends ZIndexProps,
-    TopProps<{}, SpacingUnit>,
-    RightProps<{}, SpacingUnit>,
-    BottomProps<{}, SpacingUnit>,
-    LeftProps<{}, SpacingUnit> {
-  position?: OGPositionProps["position"]
-}
